feat(client): refactor component imports for cleaner structure and add ColorSchemeToggle and FeaturesCards components with styles
This update improves the organization of component imports in `page.tsx` and introduces new `ColorSchemeToggle` and `FeaturesCards` components along with their associated styles. This enhances the user interface by providing a way to toggle color schemes and display feature cards effectively.

feat(FeaturesCards): add FeaturesCards component to showcase key features with icons and descriptions for better user engagement

feat(HeroImageRight): add new CSS module for HeroImageRight component to enhance styling and layout responsiveness

feat(client): add HeroImageRight and Welcome components for enhanced UI

This commit introduces two new components, HeroImageRight and Welcome, to improve the user interface of the application. The HeroImageRight component features a title, description, and a call-to-action button, while the Welcome component provides a welcoming message and guidance for users. Additionally, a corresponding CSS module for styling the Welcome component is included. These additions aim to enhance user engagement and provide a better onboarding experience.
