+++ apps/client/app/page.tsx
@@ -1,8 +1,8 @@
 import {FC} from 'react';
-import {ColorSchemeToggle} from '../src/components/ColorSchemeToggle/ColorSchemeToggle';
-import {FeaturesCards} from '../src/components/FeatureCards/FeaturesCards';
-import {HeroImageRight} from '../src/components/HeroImageRight/HeroImageRight';
-import {Welcome} from '../src/components/Welcome/Welcome';
+import {ColorSchemeToggle} from '../src/components/ColorSchemeToggle';
+import {FeaturesCards} from '../src/components/FeaturesCards';
+import {HeroImageRight} from '../src/components/HeroImageRight';
+import {Welcome} from '../src/components/Welcome';
 
 const Index: FC = () => {
     return (
