+++ apps/client/src/components/HeroImageRight.tsx
